Solution to Find the Sum of Digits of Each Element in an Array

+++ Find the Sum of Digits of Each Element in an Array@@676c02002dbc8c3e7a280b7f/code.c
@@ -9,7 +9,7 @@ int main(){
     }
     for(int i =0;i<n;i++){
         int sum = 0;
-        int num = arr[i]/10;
+        float num = arr[i]/10;
         sum = sum + num + arr[i]%10;
         printf("%d ",sum);
